Add unit tests for user pre-registration and account management
- Implemented fixtures for pre-registration details and user data setup.
- Created tests for user pre-registration, including creation, review, and status updates.
- Added tests for listing user pre-registrations with various filters and pagination.
- Developed search functionality tests for merged users by email, username, and group ID.
- Included sorting tests for merged user listings based on different criteria.
