+++ .github/instructions/python-tests.instructions.md
@@ -27,6 +27,15 @@ This is a multi-project monorepo with two main groups of projects in the folders
 
 - If a test file exceeds 1000 lines, split it into multiple files using descriptive name suffixes (e.g. `test_users_accounts_rest_registration.py` → `test_users_accounts_rest_registration_create.py`, `test_users_accounts_rest_registration_delete.py`, etc.).
 - If the split files share fixtures, move shared fixtures to a `conftest.py`. If the number of files warrants it, group them into a subfolder with its own `conftest.py`.
+
+### Shared test data types
+
+- Dataclasses, TypedDicts, and other types used to annotate fixtures or pass structured data between test files **must not** be imported across test modules via relative or `conftest` imports — pytest's default `prepend` import mode makes this unreliable.
+- Instead, place shared test data types in the appropriate `pytest_simcore.helpers.` module (e.g. `webserver_users.py`, `storage_utils.py`). This ensures reliable imports and encourages reuse across projects.
+- Only types that are truly local to a single test file may stay in that file.
+
+### Test file naming
+
 - **Test file names must be unique across the entire project test tree**, regardless of which subfolder they live in. Generic names like `test_list.py`, `test_search.py`, or `test_create.py` are forbidden since they can easily collide with unrelated test files elsewhere. Always keep the full original filename as a prefix for the filename, even inside a dedicated subfolder. **Folder names should be a short version of the original test filename, without the `test_` prefix**:
   - ✅ `users_accounts_rest_registration/test_users_accounts_rest_registration_search.py`
   - ❌ `test_users_accounts_rest_registration/test_users_accounts_rest_registration_search.py`
